+++ features/list-detail/components/ListDetailContent.tsx
@@ -6,7 +6,8 @@ import ContainerInformation from "@features/list-detail/components/ContainerInfo
 import ReviewSection from "@features/list-detail/components/ReviewSection";
 import FloatingBar from "@features/list-detail/components/FloatingBar";
 import testImage from "@public/image/dummyImage.png";
-import useGetGatheringDetail from "@features/list-detail/hooks/useGetGatheringDetail";
+import GatheringStatusBadge from "@features/list/GatheringStatusBadge";
+import useGetGatheringDetail from "@features/list-detail/hooks/useGatheringDetail";
 import { formatDateTime } from "@utils/dateFormatter";
 
 interface ListDetailContentProps {
@@ -29,19 +30,22 @@ export default function ListDetailContent({
   return (
     <div className="mx-auto flex max-w-screen-lg flex-col gap-8 pb-20">
       <div className="flex flex-row items-center justify-center gap-8">
-        <Image
-          src={data.image || testImage}
-          alt="모임 장소 이미지"
-          className="h-[270px] w-[468px] rounded-[24px] border-2 border-gray-200 object-cover"
-          width={468}
-          height={270}
-        />
+        <div className="relative">
+          <GatheringStatusBadge registrationEnd={data.registrationEnd} />
+          <Image
+            src={data.image || testImage}
+            alt="모임 장소 이미지"
+            className="h-[270px] w-[468px] rounded-[24px] border-2 border-gray-200 object-cover"
+            width={468}
+            height={270}
+          />
+        </div>
         <ContainerInformation
           maxCount={data.capacity}
           title={data.name}
           location={data.location}
-          date={dateString} // "ex) 2월 24일"
-          time={timeString} // "ex) 14:44"
+          date={dateString}
+          time={timeString}
         />
       </div>
       <ReviewSection />
